pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "swagcli"
version = "0.2.0"
description = "Make easy command line interfaces using swagger config"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Justin Thomas", email = "git@imthor.in" }
]
dependencies = [
    "click>=8.0.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "prompt-toolkit>=3.0.0",
    "aiohttp>=3.8.0",
    "diskcache>=5.6.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "python-multipart>=0.0.6",
    "pyjwt>=2.8.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=swagcli --cov-report=term-missing" 